'asdf'

+++ info_samdu_v2/database/migrations/2023_02_16_142907_create_educational_information_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->bigInteger('specialty_code');
             $table->string('level');
             $table->string('paymentForm');
-            $table->bigInteger('mark');
+            $table->float('mark');
             $table->timestamps();
